file reorg before moving forward
