+++ src/components/categories/CategoryDisplay.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import { remove } from '../Redux/actions';
+import { remove } from './actions';
 
 class CategoryDisplay extends Component {
