+++ src/components/categories/reducers.test.js
