+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App';
+import App from './components/app/App';
 import { Provider } from 'react-redux';
-import store from './store';
+import store from './store/store';
 import './styles/main.css';
 
 ReactDOM.render(
